+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 
 ### Unreleased
 
+### v1.0.1 (2024-07-17)
+
+* Add some blackbox testing
+* Throw an exception instead of silently overwriting if response handlers attempt to send more than one value
+  for the same header.
+
 ### v1.0.0 (2024-07-16)
 
 * First version, extracted from host project